renamed and refactored
